# specs/003-accept-terms/contracts/terms-public-v1.openapi.yaml
openapi: 3.1.0
info:
  title: API pública de términos y condiciones
  version: 1.0.0
  description: Contrato same-origin para la UI autónoma de términos.
  license: {name: Uso interno del proyecto, identifier: LicenseRef-Proprietary}
servers:
  - url: /terms-api
security:
  - bearerAuth: []
tags:
  - {name: Current, description: Documento y aceptación propios}
  - {name: Administration, description: Gestión inmutable de versiones}
  - {name: Audit, description: Evidencia de solo lectura por ámbito}
paths:
  /v1/current:
    get:
      tags: [Current]
      operationId: getCurrentTerms
      summary: Obtiene la versión aplicable y la aceptación propia
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Versión aplicable y estado del actor
          headers:
            ETag: {schema: {type: string}}
            X-Request-Id: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurrentTerms'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/acceptances:
    post:
      tags: [Current]
      operationId: acceptCurrentTerms
      summary: Acepta explícitamente la versión aplicable
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AcceptanceInput'}
      responses:
        '201':
          description: Aceptación creada
          headers:
            X-Request-Id: {$ref: '#/components/headers/RequestId'}
            Idempotency-Replayed: {schema: {type: boolean}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Acceptance'}
        '200':
          description: Replay o aceptación lógica existente
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Acceptance'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Validation'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /v1/admin/versions:
    get:
      tags: [Administration]
      operationId: listTermsVersions
      summary: Lista versiones para administración
      x-required-roles: [terms_admin]
      responses:
        '200':
          description: Versiones paginadas
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/TermsVersionSummary'}}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Administration]
      operationId: createTermsVersionDraft
      summary: Crea un borrador inmutable al publicarse
      x-required-roles: [terms_admin]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VersionDraftInput'}
      responses:
        '201':
          description: Borrador creado
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TermsVersion'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Validation'}
  /v1/admin/versions/{versionId}:
    get:
      tags: [Administration]
      operationId: getTermsVersion
      summary: Obtiene una versión exacta autorizada
      x-required-roles: [terms_admin, supervisor, auditor]
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: Versión exacta
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TermsVersion'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/admin/versions/{versionId}/schedule:
    post:
      tags: [Administration]
      operationId: scheduleTermsVersion
      summary: Programa una versión futura
      x-required-roles: [terms_admin]
      parameters:
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [effectiveAt]
              properties:
                effectiveAt: {type: string, format: date-time}
      responses:
        '200':
          description: Versión programada
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TermsVersion'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Validation'}
  /v1/admin/versions/{versionId}/withdraw:
    post:
      tags: [Administration]
      operationId: withdrawTermsVersion
      summary: Retira un borrador o versión aún no vigente
      x-required-roles: [terms_admin]
      parameters:
        - $ref: '#/components/parameters/VersionId'
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Borrador o versión futura retirada
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TermsVersion'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/audit/acceptances/search:
    post:
      tags: [Audit]
      operationId: searchTermsAcceptances
      summary: Busca aceptaciones dentro del ámbito autorizado
      x-required-roles: [supervisor, auditor]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AcceptanceSearch'}
      responses:
        '200':
          description: Evidencias del ámbito autorizado
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/Acceptance'} }
                  nextCursor: {type: [string, 'null']}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/Validation'}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: true
      schema: {type: string, format: uuid}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, format: uuid}
    VersionId:
      name: versionId
      in: path
      required: true
      schema: {type: string, format: uuid}
  headers:
    RequestId: {schema: {type: string, format: uuid}}
  schemas:
    CurrentTerms:
      type: object
      additionalProperties: false
      required: [version, acceptanceStatus]
      properties:
        version: {$ref: '#/components/schemas/TermsVersion'}
        acceptanceStatus: {type: string, enum: [PENDING, ACCEPTED]}
        acceptedAt: {type: [string, 'null'], format: date-time}
    TermsVersionSummary:
      type: object
      required: [versionId, versionCode, title, contentSha256, state]
      properties:
        versionId: {type: string, format: uuid}
        versionCode: {type: string, pattern: '^[A-Z0-9][A-Z0-9._-]{0,63}$'}
        title: {type: string, minLength: 1, maxLength: 200}
        contentSha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        state: {type: string, enum: [DRAFT, SCHEDULED, EFFECTIVE, SUPERSEDED, WITHDRAWN]}
        effectiveAt: {type: [string, 'null'], format: date-time}
    TermsVersion:
      allOf:
        - $ref: '#/components/schemas/TermsVersionSummary'
        - type: object
          required: [contentFormat, content]
          properties:
            contentFormat: {const: markdown}
            content: {type: string, minLength: 1, maxLength: 524288}
            publishedAt: {type: [string, 'null'], format: date-time}
    VersionDraftInput:
      type: object
      additionalProperties: false
      required: [versionCode, title, contentFormat, content]
      properties:
        versionCode: {type: string, pattern: '^[A-Z0-9][A-Z0-9._-]{0,63}$'}
        title: {type: string, minLength: 1, maxLength: 200}
        contentFormat: {const: markdown}
        content: {type: string, minLength: 1, maxLength: 524288}
    AcceptanceInput:
      type: object
      additionalProperties: false
      required: [versionId, contentSha256]
      properties:
        versionId: {type: string, format: uuid}
        contentSha256: {type: string, pattern: '^[a-f0-9]{64}$'}
    Acceptance:
      type: object
      additionalProperties: false
      required: [acceptanceId, versionId, versionCode, acceptedAt, contentSha256]
      properties:
        acceptanceId: {type: string, format: uuid}
        versionId: {type: string, format: uuid}
        versionCode: {type: string}
        acceptedAt: {type: string, format: date-time}
        contentSha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        actorDisplay: {type: [string, 'null'], description: Valor enmascarado y autorizado}
    AcceptanceSearch:
      type: object
      additionalProperties: false
      properties:
        actorPublicId: {type: string, maxLength: 128}
        versionCode: {type: string, maxLength: 64}
        from: {type: string, format: date-time}
        to: {type: string, format: date-time}
        cursor: {type: string}
        limit: {type: integer, minimum: 1, maximum: 100, default: 25}
    Problem:
      type: object
      required: [type, title, status, code, retryable]
      examples:
        - type: https://errors.finance2.example/terms/version-changed
          title: La versión vigente cambió
          status: 409
          detail: Vuelva a revisar el documento vigente antes de aceptar.
          code: TERMS_VERSION_CHANGED
          retryable: true
          requestId: 00000000-0000-4000-8000-000000000901
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        code: {type: string}
        retryable: {type: boolean}
        requestId: {type: string, format: uuid}
        acceptanceUrl: {type: string, pattern: '^/terms(?:/|$)'}
  responses:
    Unauthorized:
      description: Sesión ausente o inválida
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    Forbidden:
      description: Rol o ámbito insuficiente
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    NotFound:
      description: Recurso no visible o inexistente
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    Conflict:
      description: Versión cambió, replay distinto o transición incompatible
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    Validation:
      description: Entrada inválida
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
    Unavailable:
      description: Documento o persistencia no verificables
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}